# erp_system/services/inventario_service/models.py
from sqlalchemy import Column, String, Float, Integer
from sqlalchemy.dialects.postgresql import UUID
import uuid # Required for default=uuid.uuid4
from .database import Base # Relative import for local package structure

class Producto(Base):
    __tablename__ = "productos"

    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
    nombre = Column(String, index=True, nullable=False)
    descripcion = Column(String, nullable=True)
    precio_costo = Column(Float, nullable=False)
    precio_venta = Column(Float, nullable=False)
    stock_actual = Column(Integer, default=0, nullable=False)
    categoria = Column(String, nullable=True)